+++ src/main/java/org/betonquest/betonquest/database/PlayerData.java
@@ -306,7 +306,7 @@ public class PlayerData {
         if (obj == null) {
             return;
         }
-        final String data = obj.getDefaultDataInstruction();
+        final String data = obj.getDefaultDataInstruction(playerID);
         if (addRawObjective(objectiveID.toString(), data)) {
             saver.add(new Record(UpdateType.ADD_OBJECTIVES, playerID, objectiveID.toString(), data));
         }